// admin-clients/libs/cpanel/venues/feature/standard-venue-tpl-manager/src/lib/models/label-group/venue-template-label-group-type.enum.ts
export enum VenueTemplateLabelGroupType {
    state = 'state',
    blockingReason = 'blockingReason',
    priceType = 'priceType',
    quota = 'quota',
    seasonSessions = 'seasonSessions',
    sessionSeasons = 'sessionSeasons',
    visibility = 'visibility',
    accessibility = 'accessibility',
    gate = 'gate',
    seasonTicketLinkable = 'seasonTicketLinkable',
    sessionPackLink = 'sessionPackLink',
    sessionPacks = 'sessionPacks',
    firstCustomLabelGroup = 'firstCustomLabelGroup',
    secondCustomLabelGroup = 'secondCustomLabelGroup'
}